requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
PyMySQL
python-dotenv
cryptography
